Fix for issue with new version of Numpy
Upon trying to read in the test data the program throws the error "only integer scalar arrays can be converted to a scalar index".

"This appears to be an issue with the latest version of Numpy. A recent change made it an error to treat a single-element array as a scalar for the purposes of indexing." <--- https://stackoverflow.com/questions/42128830/typeerror-only-integer-scalar-arrays-can-be-converted-to-a-scalar-index

Fixed by indexing the desired value in the single scalar array.